Finish api endpoint energy monitor

+++ database/migrations/2022_01_30_232820_create_hardware_power_loads_today_table.php
@@ -82,7 +82,7 @@ class CreateHardwarePowerLoadsTodayTable extends Migration
                 ->nullable()
                 ->default(0)
                 ->comment('Amperaje máximo (A)');
-            $table->decimal('amperage', 10, 2)
+            $table->double('amperage')
                 ->nullable()
                 ->default(0)
                 ->comment('Amperaje total (Ah)');
