+++ database/migrations/2022_01_30_232824_create_hardware_power_generators_historical_table.php
@@ -46,7 +46,7 @@ class CreateHardwarePowerGeneratorsHistoricalTable extends Migration
                 ->nullable()
                 ->default(0)
                 ->comment('Número de veces que se ha cargado la batería por completo');
-            $table->decimal('amperage', 10, 2)
+            $table->double('amperage')
                 ->nullable()
                 ->default(0)
                 ->comment('Carga total en Ah que ha sido almacenado en la batería');
